% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{df_to_xml_team_metric}
\alias{df_to_xml_team_metric}
\title{Convert a dataframe to SportsCode ready XML, rows as teams with one metric}
\usage{
df_to_xml_team_metric(dat, first_half_start, second_half_start,
  start, end, team_name, metric_col_value,
  period_col, export_filename, metric_label, extra_time_start = NA,
  extra_time_second_half_start = NA)
}
\arguments{
\item{dat}{A dataframe to convert to XML, dataframe format}

\item{first_half_start}{The second value of the start of the first half within your SportsCode package. Numeric value}

\item{second_half_start}{The second value of the start of the second half within your SportsCode package. Numeric value}

\item{start}{The moment in time the event / occurrence starts. Select the column order by numeric value}

\item{end}{The moment in time the event / occurrence ends. Select the column order by numeric value}

\item{team_name}{The name of the team. Select the column order by numeric value}

\item{metric_col_value}{The metric values. Select the column order by numeric value}

\item{period_col}{The period of play, needs to be 1,2,3,4. Select the column order by numeric value}

\item{export_filename}{The filename for export. String which must end in ".xml"}

\item{metric_label}{The name of the metric to be displayed within SportsCode. Ideally ends with ": "}

\item{extra_time_start}{The second value of the start of the extra time within your SportsCode package. Numeric value. Default = NA}

\item{extra_time_second_half_start}{The second value of the start of the second half of extra time within your SportsCode package. Numeric value. Default = NA}
}
\value{
An exported XML
}
\description{
Convert a dataframe to SportsCode ready XML, rows as players with one metric
}
