% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportscoder.R
\name{read_sportscode_xml}
\alias{read_sportscode_xml}
\title{Analyse Hudl SportsCode data in R}
\usage{
read_sportscode_xml(sportscode_xml_file_path, format = "long")
}
\arguments{
\item{format}{Choose either long, tidy, matrix or matrix_group Format
"Long Format" returns a data frame with a row per code & label group combination
"Tidy Format" returns a data frame with a single row per code instance (tag) and corresponding labels as columns
"Matrix" is a summary format returning a single row per code, with counts of corresponding groups & labels (Based on the sportscode matrix)
"Matrix Group" is a summary format returning a single row per code, with counts of corresponding groups only (Based on the sportscode matrix)}

\item{path}{Path to SportsCode XML file}
}
\value{
A data.frame
}
\description{
Load SportsCode XML Edit file into a tibble for analysis
}
\examples{
read_sportscode_xml('Sportscode-edit-file.xml')
read_sportscode_xml('Sportscode-edit-file.xml', 'long')
read_sportscode_xml('Sportscode-edit-file.xml', 'tidy')
read_sportscode_xml('Sportscode-edit-file.xml', 'matrix')
}
